# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Librería API",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/resenas/{id}": {
            "get": {
                "tags": [
                    "resena-_libro-controller"
                ],
                "operationId": "getResena",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resena_libro"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "resena-_libro-controller"
                ],
                "operationId": "updateResena",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Resena_libro"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resena_libro"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "resena-_libro-controller"
                ],
                "operationId": "deleteResena",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/regiones/{id}": {
            "get": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "getRegion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Region"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "updateRegion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegionUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Region"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "deleteRegion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/usuarios/registrar": {
            "post": {
                "tags": [
                    "usuario-controller"
                ],
                "operationId": "registrarUsuario",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsuarioDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/resenas": {
            "get": {
                "tags": [
                    "resena-_libro-controller"
                ],
                "operationId": "getAllResenas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Resena_libro"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "resena-_libro-controller"
                ],
                "operationId": "createResena",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Resena_libro"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resena_libro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/regiones": {
            "get": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "getAllRegions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Region"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "createRegion",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegionCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Region"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/intercambios/libro/{id}": {
            "get": {
                "tags": [
                    "intercambio-controller"
                ],
                "operationId": "historialPorLibro",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Intercambio"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/usuarios/correo/{correo}": {
            "delete": {
                "tags": [
                    "usuario-controller"
                ],
                "operationId": "eliminarUsuarioPorCorreo",
                "parameters": [
                    {
                        "name": "correo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Autor": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Genero": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Libro": {
                "type": "object",
                "properties": {
                    "codigoBarra": {
                        "type": "string"
                    },
                    "titulo": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "anio_publicacion": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "portada_url": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "propietario": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "autores": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Autor"
                        }
                    },
                    "generos": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Genero"
                        }
                    }
                }
            },
            "Resena_libro": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "titulo_resena": {
                        "type": "string"
                    },
                    "comentario": {
                        "type": "string"
                    },
                    "calificacion": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date"
                    },
                    "libro": {
                        "$ref": "#/components/schemas/Libro"
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/Usuario"
                    }
                }
            },
            "Usuario": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "apellido": {
                        "type": "string"
                    },
                    "correo": {
                        "type": "string"
                    },
                    "contrasena": {
                        "type": "string"
                    }
                }
            },
            "RegionUpdateRequest": {
                "type": "object",
                "properties": {
                    "nombre_region": {
                        "type": "string"
                    }
                }
            },
            "Region": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
...
            "UsuarioDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "apellido": {
                        "type": "string"
                    },
                    "correo": {
                        "type": "string"
                    },
                    "contrasena": {
                        "type": "string"
                    }
                }
            },
            "RegionCreateRequest": {
                "type": "object",
                "properties": {
                    "nombre_region": {
                        "type": "string"
                    }
                }
            },
...
            "Intercambio": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "libro": {
                        "$ref": "#/components/schemas/Libro"
                    },
                    "solicitante": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "destinatario": {
                        "$ref": "#/components/schemas/Usuario"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "SOLICITADO",
                            "ACEPTADO",
                            "RECHAZADO",
                            "EN_PROGRESO",
                            "FINALIZADO",
                            "CANCELADO"
                        ]
                    },
                    "fechaSolicitud": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "fechaAceptacion": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "fechaFinalizacion": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}