# src/main/java/com/example/libreria_app/dto/RegionCreateRequest.java
package com.example.libreria_app.dto;

public class RegionCreateRequest {
    private String nombre_region;

    public RegionCreateRequest() {
    }

    public String getNombre_region() {
        return nombre_region;
    }

    public void setNombre_region(String nombre_region) {
        this.nombre_region = nombre_region;
    }
}